Switch to one file per method for kafka controller
